+++ src/main/java/com/codecool/krk/lucidmotors/queststore/models/AvailableQuest.java
@@ -3,18 +3,20 @@ package com.codecool.krk.lucidmotors.queststore.models;
 import com.codecool.krk.lucidmotors.queststore.dao.AvailableQuestDao;
 import com.codecool.krk.lucidmotors.queststore.exceptions.DaoException;
 
+import java.math.BigInteger;
+
 public class AvailableQuest extends AbstractQuest {
 
     private AvailableQuestDao availableQuestDao = AvailableQuestDao.getDao();
 
     public AvailableQuest(String name, QuestCategory questCategory, String description,
-                          Integer value) throws DaoException {
+                          BigInteger value) throws DaoException {
 
         super(name, questCategory, description, value);
     }
 
     public AvailableQuest(String name, QuestCategory questCategory, String description,
-                          Integer value, Integer id) throws DaoException {
+                          BigInteger value, Integer id) throws DaoException {
 
         super(name, questCategory, description, value, id);
     }
@@ -32,7 +34,7 @@ public class AvailableQuest extends AbstractQuest {
         Integer id = this.getId();
         String name = this.getName();
         String description = this.getDescription();
-        Integer value = this.getValue();
+        BigInteger value = this.getValue();
 
         return ("\t" + id + ". " + name + ": " + description + "; value: " + value);
     }
